Fix auto-prune timers being skipped in script mode (#6072)
